Numerous Empathing and Siphoning enhancements
Siphoners will now call out a maximum of 5 times that they’re
exhausted, rather than spamming the hell out of the ship. Also, as long
as the empath is able to cast Clarity, whenever the ship is shockwaved,
the empath will now clarity the chair so they can make their escape.

Plus, several documentation fixes and additions, such as clarifying the
README a bit and adding in the Collector documentation.